feat: implement usePaginator hook and PageResponse interface; refactor rangeToPagesIndexes for improved pagination handling
