+++ src/stories/06-cursor-pagination/table/types/Range.ts
@@ -1,0 +1,1 @@
+export type Range = [start: number, end: number];
